Plots the historical rate of the entered currencies within a specific period of time.

Plots the historical rate of the entered currencies within a specific period of time.

Usage

plot_historical(start_date, end_date, currency1, currency2)

Arguments

start_date

Entered starting date in the format specified '%YYYY-%mm-%dd'

end_date

Entered ending date in the format specified '%YYYY-%mm-%dd'

currency1

The type of based currency asked for plotting

currency2

The type of exchange currency asked for plotting

Value

A plot showing the performance of the currency.

Examples

plot_historical('2020-05-23', '2022-05-30', 'USD', 'CAD')
#> 
#> Attaching package: ‘plotly’
#> The following object is masked from ‘package:ggplot2’:
#> 
#>     last_plot
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
#> The following object is masked from ‘package:graphics’:
#> 
#>     layout